% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{aib_panel_info}
\alias{aib_panel_info}
\title{Check panel balance}
\usage{
aib_panel_info(object)
}
\arguments{
\item{object}{An \code{aibias} object.}
}
\value{
A tibble summarizing observation counts per unit.
}
\description{
Check panel balance
}
