% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.aibias.R
\name{print.aibias}
\alias{print.aibias}
\title{Print an aibias object}
\usage{
\method{print}{aibias}(x, ...)
}
\arguments{
\item{x}{An \code{aibias} object.}

\item{...}{Ignored.}
}
\description{
Print an aibias object
}
